requirements.txt
streamlit 
Pillow 
requests
